code cleanup

+++ src/ch7_methods/Calculator.java
@@ -1,22 +1,22 @@
 package ch7_methods;
 
 public class Calculator {
-    public static int add(int a, int b) {
+    public static float add(float a, float b) {
         System.out.println("I will add " + a + " and " + b);
         return a + b;
     }
 
-    public static int subtract(int a, int b) {
+    public static float subtract(float a, float b) {
         System.out.println("I will subtract " + a + " and " + b);
         return a - b;
     }
 
-    public static int multiply(int a, int b) {
+    public static float multiply(float a, float b) {
         System.out.println("I will multiply " + a + " and " + b);
         return a * b;
     }
 
-    public static int divide(int a, int b) {
+    public static float divide(float a, float b) {
         System.out.println("I will divide " + a + " and " + b);
         return a / b;
     }
